+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # MountsJournal
 
-* Fixed restoration of filters.
-* Fixed sorting when set/unset a favorite if the journal is hidden.
+* Fixed alpha of custom animation panel.